real `report.conf` loader (#361)
## Context & Scope
We have more than 1 test now that need to check the file `report.conf`.
It is time to have a real loader of this file in `test/utils` so that
various tests can use it

## Testing
Already tested in `test_coreneuron_save_restore.py`. If you want I can
add a dedicated test in `test_test_utils.py`. I do not know where to
plop the `report.conf`... maybe I can create it on the fly...

Let me know if you prefer this option @jamesgking