`compartment sets`: stimulus (#378)
## Context
Fix: #368

## Scope

Introduce the possibility to use compartment sets even in the input
blocks

- add `get_point_list` to the stimuls manager. Now we do not pass the
target directly to the stimulus object. We pass directly the points. In
this way we can have only one function to pick the points based if
compartment_sets is there or not
- add a `get_compartment_set` to the `target_manager`
- improve `get_local_gids` to fully leverage `libsonata.Selection`.
`SelectionNodeSet.intersection` has been modified accordingly to return
a `libsonata.Selection`
- unification of `get_gids` and `get_final_gids` in `NodeSetTarget`. It
gets the same treatment as `SelectionNodeSet`. Marginal speedup due to
using `libsonata.Selection` for merging ranges.
- rename `TPointList` to `TargetPointList`
- `_tpoints` to `_target_points`

Note: if `compartment_sets` is there the `node_sets` is empty (enforced
by libsonata). This makes the target manager always pick the `_ALL_`
target. from there, we select the population provided by the
`compartment_sets` and proceed to pick the points from there. I hope
this is a sensible way of picking the points. I am open to suggestions
if you think this may be dangerous

## Testing

Tested only for the pulse input. However, given how it is implemented,
additional tests may be superfluous. Let me know if you want more

- new test added: `tests/unit/test_compartment_set_stimulus.py`