make neurodamus 0-based (again?) (#391)
## Context

fix #390 

## Scope

remove all the +-1s.

Note: in order to maintain reproducibility for everything that uses a stimulus that is gid-sensitive (a stimulus that uses an rng generator that takes the gid as a seed), I changed these seeds to take the gid + 1 instead. I added a note to expain the +1.

fix `test_conn_manager_syn_stats`. It was expecting 1-based ids while we provide 0-based ones. Results change



## Testing

- all the tests are passing with the old traces
- `check_signal_peaks` threshold in `test_spont_minis_simple` has been lowered to 0.3 because I noticed that it was not getting all the peaks. Anyway, with the old 0.5 the peaks were the same
- added #417 to investigate why my typo did not affect the tests